Fix #3: Handle systems with > MAXVIFS (32) interfaces
The UNIX multicast routing stack is by default hard-coded to a maximum
of 32 virtual interfaces (VIFs) that it can route between.  Modern
systems today often have (much) more physical interfaces than that so
we must check that we do not go out of bounds on the internal array.

We must also make sure there is enough room for the register VIF, i.e.
the tunnel to the Rendez-Vous Point.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>