Fix broken PIM checksum on interfaces with multiple link-local addresses
When an interface has more than one link-local address then this can
lead to a broken checksum in multicasted PIM messages:

This happens when we choose one link-local address for the IPv6 pseudo
header and PIM checksum calculation but the kernel chooses the other one
for the source address of the packet when transmitting.

For non-multicasted PIM packets the source address is already set
explicitly to the one the checksum was calculated over. With this patch
we do the same for multicasted ones to solve this broken checksum issue.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>