Fix call to SIOCGETSGCNT_IN6, must use correct socket
The call to read the kernel (S,G) count, for detecting stale inbound
interfaces, used the wrong socket.  On Linux the same socket as MFC
entries are added on (the MRT_INIT socket) must be used also for the
stats of said MFC entries.

Since this call failed pim6sd reverted back to trying to delete the
route for each ageout interval.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>