Fix #26: fix segfault in make_mld6v2_msg()
This is basically a sync with the make_mld6_msg() code, which is very
similar, but used calloc() instead of malloc() to ensure the cmsghdr
data is actually cleared properly.

This seems to have been the root cause of the segfault rather than
what was proposed in the GitHub issue.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>