Fix MLDv2 report reception by joining the all-mld2-routers group
On Linux 6.17 when configuring an interface like
`phyint veth0 mld_version any enable;` in pim6sd then pim6sd will
correctly send an MLDv2 query, which a host will reply to with
an according MLDv2 report. However pim6sd would not receive this
MLDv2 report on its SOCK_RAW/IPPROTO_ICMPV6 socket.

For this socket to receive an MLDv2 report we need to join the
all-mldv2-capable routers multicast address, which is also the IPv6
destination address of such a report. With this change at least the
"Reported MLD Group" table in pim6stat could successfully be populated
with an SSM entry.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>