Introduce source_outgoing_interface config option
As described in the additions to the README in this commit and in more
detail on this dn42 wiki page linked below it is important that the
unicast routing table used by pim6sd does not contain excessive paths.
Unicast-only routers incapable of PIM-SD can lead to black hole
routing.

The plausible option to avoid this would be to add a config option to
pim6sd to select an alternative routing table (as already tried by
Marek Küthe). But unfortunately, at least on Linux the rtm_table
attribute does not seem to be implemented for the RTM_GETROUTE netlink
request yet. Also the "ip -6 route get" command has no option to specify
a table id, probably for the same reason. Adding the NLM_F_ROOT|NLM_F_MATCH
flags sounds as if it were supposed to go in the correct direction for an
RTM_GETROUTE, but at least the netlink manpage says that NLM_F_MATCH
is not implemented yet. Tests seemed to confirm this.

Therefore this commit adds an alternative option, a small detour:
It allows using a VRF, which is assigned to a routing table, instead.
The newly added "source_outgoing_interface" behaves similar to
adding the "oif" or "vrf" option to an "ip -6 route get" command.
Which for a VRF will then make pim6sd use the unicast routing table
attached to the VRF in its unicast route lookups.

Closes #41.

Link: https://wiki.dn42.dev/howto/pim-multicast-vs-unicast-issue

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>