autoconf: fix cross compilation of netlink check
AC_TRY_RUN() returns an error by default on cross compilation - unless
an alternative action is given. For the netlink test such an alternative
action is currently missing, which in turn makes an OpenWrt build fail,
for instance.

For cross compiled binaries it does not make sense to runtime test the
netlink interface against the compiling host. Therefore only do a compile
time test to check for the availability of netlink when cross-compiling.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>